Making sure that x1 and y1 are floats and not numpy.float64 which cannot be marshaled by the xml rpc server

+++ mxcubecore/HardwareObjects/Shapes.py
@@ -417,8 +417,8 @@ class Grid(Shape):
         d["motor_positions"] = self.cp_list[0].as_dict()
 
         # MXCuBE - 2 WF compatability
-        d["x1"] = (self.beam_pos[0] - d["screen_coord"][0]) / self.pixels_per_mm[0]
-        d["y1"] = (self.beam_pos[1] - d["screen_coord"][1]) / self.pixels_per_mm[1]
+        d["x1"] = float((self.beam_pos[0] - d["screen_coord"][0]) / self.pixels_per_mm[0])
+        d["y1"] = float((self.beam_pos[1] - d["screen_coord"][1]) / self.pixels_per_mm[1])
         d["steps_x"] = d["num_cols"]
         d["steps_y"] = d["num_rows"]
         d["dx_mm"] = d["width"] / self.pixels_per_mm[0]
@@ -426,4 +426,5 @@ class Grid(Shape):
         d["beam_width"] = d["cell_width"]
         d["beam_height"] = d["cell_height"]
         d["angle"] = 0
+
         return d
